+++ app/src/main/java/com/example/aibudgetapp/ui/screens/budget/BudgetViewModel.kt
@@ -45,7 +45,7 @@ class BudgetViewModel(
         name: String,
         chosentype: String,
         chosencategory: String,
-        amount: Int,
+        amount: Double,
         checked: Boolean,
         startDate: String?,
         endDate: String?
